move serveSignedManifest method into Server module
