+++ packages/expo-updates/e2e/__tests__/tests/basic.ts
@@ -1,7 +1,4 @@
-import crypto from 'crypto';
-import fs from 'fs';
 import path from 'path';
-import { Dictionary, serializeDictionary } from 'structured-headers';
 import { setTimeout } from 'timers/promises';
 import uuid from 'uuid/v4';
 
@@ -25,39 +22,6 @@ if (!repoRoot) {
 
 const projectRoot = process.env.TEST_PROJECT_ROOT ?? path.resolve(repoRoot, '..', 'updates-e2e');
 const updateDistPath = path.join(projectRoot, 'dist');
-const codeSigningPrivateKeyPath = path.join(projectRoot, 'keys', 'private-key.pem');
-
-async function getPrivateKeyAsync() {
-  const pemBuffer = fs.readFileSync(path.resolve(codeSigningPrivateKeyPath));
-  return pemBuffer.toString('utf8');
-}
-
-function signRSASHA256(data: string, privateKey: string) {
-  const sign = crypto.createSign('RSA-SHA256');
-  sign.update(data, 'utf8');
-  sign.end();
-  return sign.sign(privateKey, 'base64');
-}
-
-function convertToDictionaryItemsRepresentation(obj: { [key: string]: string }): Dictionary {
-  return new Map(
-    Object.entries(obj).map(([k, v]) => {
-      return [k, [v, new Map()]];
-    })
-  );
-}
-
-async function serveUpdateWithManifest(manifest: any) {
-  const privateKey = await getPrivateKeyAsync();
-  const manifestString = JSON.stringify(manifest);
-  const hashSignature = signRSASHA256(manifestString, privateKey);
-  const dictionary = convertToDictionaryItemsRepresentation({
-    sig: hashSignature,
-    keyid: 'main',
-  });
-  const signature = serializeDictionary(dictionary);
-  Server.serveManifest(manifest, { 'expo-protocol-version': '0', 'expo-signature': signature });
-}
 
 export default () =>
   describe('Basic e2e', () => {
@@ -119,7 +83,7 @@ export default () =>
       };
 
       Server.start(SERVER_PORT);
-      await serveUpdateWithManifest(manifest);
+      await Server.serveSignedManifest(manifest, projectRoot);
       await Simulator.installApp();
       await Simulator.startApp();
       const firstRequest = await Server.waitForUpdateRequest(10000 * TIMEOUT_BIAS);
@@ -167,7 +131,7 @@ export default () =>
       };
 
       Server.start(SERVER_PORT);
-      await serveUpdateWithManifest(manifest);
+      await Server.serveSignedManifest(manifest, projectRoot);
       await Simulator.installApp();
       await Simulator.startApp();
       await Server.waitForUpdateRequest(10000 * TIMEOUT_BIAS);
@@ -227,7 +191,7 @@ export default () =>
       };
 
       Server.start(SERVER_PORT);
-      await serveUpdateWithManifest(manifest);
+      await Server.serveSignedManifest(manifest, projectRoot);
       await Simulator.installApp();
       await Simulator.startApp();
       const response = await Server.waitForResponse(10000 * TIMEOUT_BIAS);
@@ -269,7 +233,7 @@ export default () =>
       };
 
       Server.start(SERVER_PORT);
-      await serveUpdateWithManifest(manifest);
+      await Server.serveSignedManifest(manifest, projectRoot);
       await Simulator.installApp();
       await Simulator.startApp();
       await Server.waitForUpdateRequest(10000 * TIMEOUT_BIAS);
