+++ packages/expo-updates/e2e/__tests__/utils/server.ts
@@ -1,5 +1,8 @@
+import crypto from 'crypto';
 import express from 'express';
+import fs from 'fs';
 import path from 'path';
+import { Dictionary, serializeDictionary } from 'structured-headers';
 import { setTimeout } from 'timers/promises';
 
 const app: any = express();
@@ -76,11 +79,6 @@ app.get('/update', (req: any, res: any) => {
   }
 });
 
-export function serveManifest(manifest: any, headers: any = null) {
-  manifestToServe = manifest;
-  manifestHeadersToServe = headers;
-}
-
 export async function waitForUpdateRequest(timeout: number) {
   const finishTime = new Date().getTime() + timeout;
   while (!updateRequest) {
@@ -94,4 +92,42 @@ export async function waitForUpdateRequest(timeout: number) {
   const request = updateRequest;
   updateRequest = null;
   return request;
+}
+
+export function serveManifest(manifest: any, headers: any = null) {
+  manifestToServe = manifest;
+  manifestHeadersToServe = headers;
+}
+
+async function getPrivateKeyAsync(projectRoot: string) {
+  const codeSigningPrivateKeyPath = path.join(projectRoot, 'keys', 'private-key.pem');
+  const pemBuffer = fs.readFileSync(path.resolve(codeSigningPrivateKeyPath));
+  return pemBuffer.toString('utf8');
+}
+
+function signRSASHA256(data: string, privateKey: string) {
+  const sign = crypto.createSign('RSA-SHA256');
+  sign.update(data, 'utf8');
+  sign.end();
+  return sign.sign(privateKey, 'base64');
+}
+
+function convertToDictionaryItemsRepresentation(obj: { [key: string]: string }): Dictionary {
+  return new Map(
+    Object.entries(obj).map(([k, v]) => {
+      return [k, [v, new Map()]];
+    })
+  );
+}
+
+export async function serveSignedManifest(manifest: any, projectRoot: string) {
+  const privateKey = await getPrivateKeyAsync(projectRoot);
+  const manifestString = JSON.stringify(manifest);
+  const hashSignature = signRSASHA256(manifestString, privateKey);
+  const dictionary = convertToDictionaryItemsRepresentation({
+    sig: hashSignature,
+    keyid: 'main',
+  });
+  const signature = serializeDictionary(dictionary);
+  serveManifest(manifest, { 'expo-protocol-version': '0', 'expo-signature': signature });
 }
